moved connector map simplified test in connector type pipe created constant for select all and use it cut pricingDefinitions pipe in two and adjust call to each simplified formatter simplified context set (everything in tabledatatsource)
