+++ src/app/shared/formatters/app-connector-type-selection.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { CONNECTOR_ALL_TYPES_MAP } from '../../types/Connector';
+import { CONNECTOR_ALL_TYPES_MAP } from '../../shared/model/charging-stations.model';
 
 export const CONNECTOR_TYPE_SELECTION_MAP = [...CONNECTOR_ALL_TYPES_MAP.values()].filter((item: any) => item.key !== 'U');
 
@@ -8,13 +8,12 @@ export const CONNECTOR_TYPE_SELECTION_MAP = [...CONNECTOR_ALL_TYPES_MAP.values()
 export class AppConnectorTypeSelectionPipe implements PipeTransform {
   public transform(type: string, target: string = 'icon'): any {
     // Return the found key
-    const foundConnectorType = CONNECTOR_TYPE_SELECTION_MAP.find(
-      (connectorType) => connectorType.key === type);
+    const foundConnectorType = CONNECTOR_TYPE_SELECTION_MAP.find((connectorType) => connectorType.key === type);
     if (target === 'icon') {
-      return (foundConnectorType ? foundConnectorType.svgIconName : 'U');
+      return foundConnectorType.svgIconName;
     }
     if (target === 'text') {
-      return (foundConnectorType ? foundConnectorType.description : 'chargers.connector_type_unknown');
+      return foundConnectorType.description;
     }
   }
 }
